requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.1
django-environ==0.11.2
django-googledrive-storage==1.6.0
django-import-export==3.3.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
google-api-core==2.10.2
google-api-python-client==2.114.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
protobuf==4.25.2
psycopg==3.1.17
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
